Add reference links to address/parsing utils, refactor E64Number.ts to io.ts
