+++ src/identity/actions.ts
@@ -9,7 +9,7 @@ import {
 } from 'src/identity/reducer'
 import { ImportContactsStatus } from 'src/identity/types'
 import { Recipient } from 'src/recipients/recipient'
-import { type E164Number } from 'src/utils/E164Number'
+import { type E164Number } from 'src/utils/io'
 
 export enum Actions {
   SET_SEEN_VERIFICATION_NUX = 'IDENTITY/SET_SEEN_VERIFICATION_NUX',
